Update SelectionGroupInspector.cs
Updated class indentation.

Moved fields to top of class.

Updated field naming to be consistent with Unity's naming style.

Moved OnEnable() method to top for consistency with other scripts and so the methods are in call order which can make the code easier to read.

Updated curly braces to be on their own line for consistancy.